move escape functions to its own file.
remove redundancy.
