+++ packages/driver/src/dom/elements/utils.ts
@@ -5,7 +5,6 @@ import $window from '../window'
 import $document from '../document'
 
 const whitespaces = /\s+/g
-const quotesRe = /('|")/g
 
 // When multiple space characters are considered as a single whitespace in all tags except <pre>.
 export const normalizeWhitespaces = (elem) => {
@@ -27,12 +26,6 @@ export const isSame = function ($el1, $el2) {
 
 export const isSelector = ($el: JQuery<HTMLElement>, selector) => {
   return $el.is(selector)
-}
-
-export function escapeQuotes (text) {
-  // convert to str and escape any single
-  // or double quotes
-  return (`${text}`).replace(quotesRe, '\\$1')
 }
 
 export function switchCase (value, casesObj, defaultKey = 'default') {
